moved stdlib to separate file; added exit function
