+++ src/main.c
@@ -19,212 +19,12 @@
 #include "pico/analysis/typecheck.h"
 #include "pico/codegen/codegen.h"
 #include "pico/eval/call.h"
+#include "pico/values/stdlib.h"
 #include "pico/values/types.h"
 
 typedef struct repl_opts {
     bool debug_print;
 } repl_opts;
-
-PiType mk_binop_type(Allocator* a, PrimType a1, PrimType a2, PrimType r) {
-    PiType* i1 = mem_alloc(sizeof(PiType), a);
-    PiType* i2 = mem_alloc(sizeof(PiType), a);
-    PiType* i3 = mem_alloc(sizeof(PiType), a);
-
-    i1->sort = TPrim;
-    i1->prim = a1;
-    i2->sort = TPrim;
-    i2->prim = a2;
-    i3->sort = TPrim;
-    i3->prim = r;
-
-    PiType type;
-    type.sort = TProc;
-    PtrArray args = mk_ptr_array(2, a);
-    push_ptr(i1, &args);
-    push_ptr(i2, &args);
-
-    type.proc.args = args;
-    type.proc.ret = i3;
-
-    return type;
-}
-
-void build_binary_fun(Assembler* ass, BinaryOp op, Allocator* a, ErrorPoint* point) {
-    build_unary_op (ass, Pop, reg(RCX), a, point);
-    build_unary_op (ass, Pop, reg(RBX), a, point);
-    build_unary_op (ass, Pop, reg(RAX), a, point);
-    build_binary_op (ass, op, reg(RAX), reg(RBX), a, point);
-    build_unary_op (ass, Push, reg(RAX), a, point);
-    build_unary_op (ass, Push, reg(RCX), a, point);
-    build_nullary_op (ass, Ret, a, point);
-}
-
-void build_comp_fun(Assembler* ass, UnaryOp op, Allocator* a, ErrorPoint* point) {
-    build_unary_op (ass, Pop, reg(RCX), a, point);
-    build_unary_op (ass, Pop, reg(RBX), a, point);
-    build_unary_op (ass, Pop, reg(RAX), a, point);
-    build_binary_op (ass, Cmp, reg(RAX), reg(RBX), a, point);
-    build_unary_op (ass, op, reg(RAX), a, point);
-    build_unary_op (ass, Push, reg(RAX), a, point);
-    build_unary_op (ass, Push, reg(RCX), a, point);
-    build_nullary_op (ass, Ret, a, point);
-}
-
-Module* base_module(Assembler* ass, Allocator* a) {
-    Module* module = mk_module(a);
-    Symbol sym;
-
-    PiType type;
-    PiType type_val;
-    PiType* type_data = &type_val;
-    type = (PiType) {
-        .sort = TKind,
-        .kind.nargs = 0,
-    };
-    ErrorPoint point;
-    if (catch_error(point)) {
-        panic(point.error_message);
-    }
-
-    type_val = mk_prim_type(Unit);
-    sym = string_to_symbol(mv_string("Unit"));
-    add_def(module, sym, type, &type_data);
-
-    type_val = mk_prim_type(Bool);
-    sym = string_to_symbol(mv_string("Bool"));
-    add_def(module, sym, type, &type_data);
-
-    type_val = mk_prim_type(Address);
-    sym = string_to_symbol(mv_string("Address"));
-    add_def(module, sym, type, &type_data);
-
-    type_val = mk_prim_type(Int_64);
-    sym = string_to_symbol(mv_string("I64"));
-    add_def(module, sym, type, &type_data);
-
-    type_val = mk_prim_type(Int_32);
-    sym = string_to_symbol(mv_string("I32"));
-    add_def(module, sym, type, &type_data);
-
-    type_val = mk_prim_type(Int_16);
-    sym = string_to_symbol(mv_string("I16"));
-    add_def(module, sym, type, &type_data);
-
-    type_val = mk_prim_type(Int_8);
-    sym = string_to_symbol(mv_string("I8"));
-    add_def(module, sym, type, &type_data);
-
-    build_binary_fun(ass, Add, a, &point);
-    type = mk_binop_type(a, Int_64, Int_64, Int_64);
-    sym = string_to_symbol(mv_string("+"));
-    add_fn_def(module, sym, type, ass, NULL);
-    clear_assembler(ass);
-
-    build_binary_fun(ass, Sub, a, &point);
-    sym = string_to_symbol(mv_string("-"));
-    add_fn_def(module, sym, type, ass, NULL);
-    clear_assembler(ass);
-    delete_pi_type(type, a);
-
-    build_comp_fun(ass, SetL, a, &point);
-    type = mk_binop_type(a, Int_64, Int_64, Bool);
-    sym = string_to_symbol(mv_string("<"));
-    add_fn_def(module, sym, type, ass, NULL);
-    clear_assembler(ass);
-
-    build_comp_fun(ass, SetG, a, &point);
-    sym = string_to_symbol(mv_string(">"));
-    add_fn_def(module, sym, type, ass, NULL);
-    clear_assembler(ass);
-
-    build_comp_fun(ass, SetE, a, &point);
-    sym = string_to_symbol(mv_string("="));
-    add_fn_def(module, sym, type, ass, NULL);
-    clear_assembler(ass);
-    delete_pi_type(type, a);
-
-    // TODO: we use int64_t as it has the requisite size (8 bytes)
-    // for pico values: currently don't support non-64 bit values 
-    int64_t former;
-    //TermFormer former;
-    type.sort = TPrim;
-    type.prim = TFormer;
-
-    former = FDefine;
-    sym = string_to_symbol(mv_string("def"));
-    add_def(module, sym, type, &former);
-
-    former = FDefine;
-    sym = string_to_symbol(mv_string("declare"));
-    add_def(module, sym, type, &former);
-
-    former = FProcedure;
-    sym = string_to_symbol(mv_string("proc"));
-    add_def(module, sym, type, &former);
-
-    former = FAll;
-    sym = string_to_symbol(mv_string("all"));
-    add_def(module, sym, type, &former);
-
-    former = FApplication;
-    sym = string_to_symbol(mv_string("$"));
-    add_def(module, sym, type, &former);
-
-    former = FProjector;
-    sym = string_to_symbol(mv_string("."));
-    add_def(module, sym, type, &former);
-
-    former = FStructure;
-    sym = string_to_symbol(mv_string("struct"));
-    add_def(module, sym, type, &former);
-
-    former = FVariant;
-    sym = string_to_symbol(mv_string(":"));
-    add_def(module, sym, type, &former);
-
-    former = FMatch;
-    sym = string_to_symbol(mv_string("match"));
-    add_def(module, sym, type, &former);
-
-    former = FIf;
-    sym = string_to_symbol(mv_string("if"));
-    add_def(module, sym, type, &former);
-
-    former = FLabels;
-    sym = string_to_symbol(mv_string("labels"));
-    add_def(module, sym, type, &former);
-
-    former = FSequence;
-    sym = string_to_symbol(mv_string("seq"));
-    add_def(module, sym, type, &former);
-
-    former = FLet;
-    sym = string_to_symbol(mv_string("let"));
-    add_def(module, sym, type, &former);
-
-    former = FIs;
-    sym = string_to_symbol(mv_string("is"));
-    add_def(module, sym, type, &former);
-
-    // Types 
-    former = FProcType;
-    sym = string_to_symbol(mv_string("Proc"));
-    add_def(module, sym, type, &former);
-
-    former = FStructType;
-    sym = string_to_symbol(mv_string("Struct"));
-    add_def(module, sym, type, &former);
-
-    former = FEnumType;
-    sym = string_to_symbol(mv_string("Enum"));
-    add_def(module, sym, type, &former);
-
-    former = FAllType;
-    sym = string_to_symbol(mv_string("All"));
-    add_def(module, sym, type, &former);
-
-    return module;
-}
 
 bool repl_iter(IStream* cin, OStream* cout, Allocator* a, Assembler* ass, Module* module, repl_opts opts) {
     // TODO (TAGS: UB BUG INVESTIGATE): Possibly need to add volatile qualifier to arena?
@@ -234,6 +34,10 @@ bool repl_iter(IStream* cin, OStream* cout, Allocator* a, Assembler* ass, Module
 
     clear_assembler(ass);
     Environment* env = env_from_module(module, &arena);
+
+    jmp_buf exit_point;
+    if (setjmp(exit_point)) goto on_exit;
+    set_exit_callback(&exit_point);
 
     ErrorPoint point;
     if (catch_error(point)) goto on_error;
@@ -320,6 +124,9 @@ bool repl_iter(IStream* cin, OStream* cout, Allocator* a, Assembler* ass, Module
  on_error:
     write_string(point.error_message, cout);
     write_string(mv_string("\n"), cout);
+    release_arena_allocator(arena);
+    return true;
+ on_exit:
     release_arena_allocator(arena);
     return false;
 }
